Report missing user id errors when loading user data
This happens if users are removed or disappear for some reason. We don't
show this in the UI quite yet (it's tricky since there's no data and
it's not really resolvable manually) but we may do in future. AFAIK
there's no way to get into this state without me manually deleting a
user account that's in a team, but it's good to report it just in case.