Take direct control of checkout pricing from Paddle
Previously this lived within Paddle, and we just redirected. Now we set
prices, we show our own prices, and we tell Paddle the specific relevant
price at the moment the checkout is requested.

This should mean a) we can configure pricing for countries that Paddle
doesn't support (most of them), b) we can version control pricing, c)
checkout pricing is independent of the currency the customer actually
pays in, d) we're a little more flexible to switch payment providers as
necessary or for specific customer groups.