Allow (explicitly) not setting an email for redirect-to-checkout API
Goal here is just to catch any cases where we accidentally don't set an
email - in general if the email isn't right in the checkout we won't
successfully link the resulting webhook to the user's account, so it's
bad.

We do want to allow it anyway though, for manual cases where we send a
checkout link, in which case it's inconvenient to need a unique link for
each user, and it's OK to just give the user instructions (and manually
fix any issues) instead.