Improve Paddle checkout handling of unsupported country codes
Paddle has issues with the DRC specifically (not for any specific reason
as far as I can tell) but this should allow users to work around it, or
at least make it clearer who is at fault.

I'm going to switch the DRC over to PayPro shortly, which should
mitigate this hopefully, but this fix will also help for any other
countries that find themselves having similar issues.