Include subscription_quantity in team app data to work around UI bug
As part of the migration to the new accounts package, the logic changed
slightly to read the quantity where available, and unfortunately if
that's not available the UI considers you as not logged in! This is
problematic.

We _probably_ shouldn't expose the subscription quantity to team
members, really (they don't need it at least) so this may be removed
from the API in future, but we can ship it for now and fix the UI first.