Make auth0-lock an explicit dependency, to fix build issues in the UI
Previously it was a dev-dep, but the direct path dependency made those
available to the accounts UI during the build process, and the main UI
has its own dependency, so that didn't cause issues. It's better to be
explicit though, since moving to proper dependencies breaks things, so
this adds that directly.