Use exponential backoff for Profitwell retries
This is useful because Paddle is very slow sometimes. Worst case ~3
hours from what I've seen recently. As long as the server runs long
enough, this should cover that. For most cases it should be a fair bit
quicker (and it's never the end of the world if we don't successfully
populate these fields - it's just extra metadata).